move qiskit circuit serialization from `css` to `gss`
